Adding negative numbers

+++ code/korotaev-pavel/src/main/java/ru/unn/agile/HexBinOctCalculator/Model/HexBinOctCalculator.java
@@ -20,32 +20,32 @@ public final class HexBinOctCalculator {
 
     public static String add(final String firstNumber, final String secondNumber,
                              final NumeralSystem system) {
-        int left = Integer.parseInt(firstNumber, system.getValue());
-        int right = Integer.parseInt(secondNumber, system.getValue());
+        int left = (int) Long.parseLong(firstNumber, system.getValue());
+        int right = (int) Long.parseLong(secondNumber, system.getValue());
         int sum = left + right;
         return convertResult(sum, system);
     }
 
     public static String subtract(final String firstNumber, final String secondNumber,
                                   final NumeralSystem system) {
-        int left = Integer.parseInt(firstNumber, system.getValue());
-        int right = Integer.parseInt(secondNumber, system.getValue());
+        int left = (int) Long.parseLong(firstNumber, system.getValue());
+        int right = (int) Long.parseLong(secondNumber, system.getValue());
         int diff = left - right;
         return convertResult(diff, system);
     }
 
     public static String multiply(final String firstNumber, final String secondNumber,
                                   final NumeralSystem system) {
-        int left = Integer.parseInt(firstNumber, system.getValue());
-        int right = Integer.parseInt(secondNumber, system.getValue());
+        int left = (int) Long.parseLong(firstNumber, system.getValue());
+        int right = (int) Long.parseLong(secondNumber, system.getValue());
         int prod = left * right;
         return convertResult(prod, system);
     }
 
     public static String divide(final String firstNumber, final String secondNumber,
                                 final NumeralSystem system) {
-        int left = Integer.parseInt(firstNumber, system.getValue());
-        int right = Integer.parseInt(secondNumber, system.getValue());
+        int left = (int) Long.parseLong(firstNumber, system.getValue());
+        int right = (int) Long.parseLong(secondNumber, system.getValue());
         if (right == 0) {
             throw new IllegalArgumentException("Division by zero!");
         } else {
